# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para sugestões de presentes personalizadas.",
        "title": "Gift App API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/profiles/{friendId}/suggestions": {
            "post": {
                "description": "Exemplo de payload: {\"occasionDetails\":\"Aniversario em 2026-08-15, prefere experiencias ao ar livre\",\"reminderID\":\"d8c8efdf-c52f-4d6b-8e2e-b83f78de4f77\"}.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "suggestion-agent"
                ],
                "summary": "Criar sugestão inicial para friend",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do friend",
                        "name": "friendId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados da ocasião",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.SuggestionCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/suggestions/agent/chat": {
            "post": {
                "description": "Exemplo de payload: {\"giftID\":\"77a6f2b8-78f4-4cc3-bcc8-1e7f2499abf0\",\"message\":\"Pode sugerir uma opcao mais economica?\"}.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "suggestion-agent"
                ],
                "summary": "Conversar para refinar sugestão por gift",
                "parameters": [
                    {
                        "description": "Payload do chat",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.SuggestionChatRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/suggestions/agent/finalize": {
            "post": {
                "description": "Exemplo de payload: {\"giftID\":\"77a6f2b8-78f4-4cc3-bcc8-1e7f2499abf0\"}.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "suggestion-agent"
                ],
                "summary": "Finalizar refinamento de sugestão por gift",
                "parameters": [
                    {
                        "description": "Payload de finalização",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.SuggestionFinalizeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "http.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "birthDate must be in format YYYY-MM-DD"
                }
            }
        },
...
        "http.SuggestionChatRequest": {
            "type": "object",
            "properties": {
                "giftID": {
                    "type": "string",
                    "example": "77a6f2b8-78f4-4cc3-bcc8-1e7f2499abf0"
                },
                "message": {
                    "type": "string",
                    "example": "Pode sugerir uma opcao mais economica?"
                }
            }
        },
        "http.SuggestionCreateRequest": {
            "type": "object",
            "properties": {
                "occasionDetails": {
                    "type": "string",
                    "example": "Aniversario em 2026-08-15, prefere experiencias ao ar livre"
                },
                "reminderID": {
                    "type": "string",
                    "example": "d8c8efdf-c52f-4d6b-8e2e-b83f78de4f77"
                }
            }
        },
        "http.SuggestionFinalizeRequest": {
            "type": "object",
            "properties": {
                "giftID": {
                    "type": "string",
                    "example": "77a6f2b8-78f4-4cc3-bcc8-1e7f2499abf0"
                }
            }
        },
...
    }
}